Unificando o código para criar o executável
