+++ .idea/workspace.xml
@@ -4,7 +4,10 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="29bb0385-53f5-40ff-bf50-0531fa48f397" name="Changes" comment="" />
+    <list default="true" id="29bb0385-53f5-40ff-bf50-0531fa48f397" name="Changes" comment="">
+      <change beforePath="$PROJECT_DIR$/codigo_python/codigo_1.py" beforeDir="false" afterPath="$PROJECT_DIR$/codigo_python/codigo_1.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/codigo_python/functions.py" beforeDir="false" afterPath="$PROJECT_DIR$/codigo_python/functions.py" afterDir="false" />
+    </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
